pyproject.toml
[tool.poetry]
name = "backend sul turismo"
version = "0.1.0"
description = ""
authors = ["João"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
